+++ src/10start.js
@@ -5,6 +5,7 @@
 // Version: 0.0.6
 // (ñ) 2014, Andrey Gershun
 //
+
 /*
 The MIT License (MIT)